Create geometry interface for a straight line
