+++ linesman/geometry/__init__.py
@@ -1,0 +1,3 @@
+from .abstract import Vector, StraightLine
+from .planar import Line
+from .spherical import Geodesic
