+++ linesman/geometry/abstract.py
@@ -1,4 +1,5 @@
 import math
+import abc
 
 
 # Floating point operations are prone to rounding error. For equality operations,
@@ -7,7 +8,7 @@ EPSILON = 10E-12
 
 
 class Vector:
-    """A vector (point) in 2-dimensional space. Instances are immutable."""
+    """A vector (point) with two coordinates."""
     def __init__(self, x, y):
         self._x = x
         self._y = y
@@ -69,77 +70,55 @@ class Vector:
         return str(self)
 
 
-class Line:
-    """Class for a line in a 2-dimensional space. Instances are immutable."""
+class StraightLine(abc.ABC):
+    """
+    Abstract base class for the line through two points that is the shortest
+    path between both points.
+    There are implementations with planar and spherical geometry. Instances of
+    the subclasses should not be mixed in use. However, the interface makes
+    switching between both interpretations easier.
+    """
     def __init__(self, a: Vector, b: Vector):
         if a == b:
-            raise ValueError('line cannot be defined by one point')
+            raise ValueError('Line cannot be defined by one point')
         self._p1 = a
         self._p2 = b
 
     def __contains__(self, other: Vector):
         """:return: whether the line contains a certain point"""
-        other = other - self._p1
-        return self.direction.parallel_to(other)
+        raise NotImplementedError
 
     def __eq__(self, other):
         """:return: whether all points described by self are also on other"""
         if type(self) != type(other):
             return NotImplemented
-        if self.direction.parallel_to(other.direction):
-            print('are parallel')
-            if self.point(0) in other:
-                return True
-            print('but no matching point')
-        return False
+        return other.point(0) in self and other.point(1) in self
 
     def __repr__(self):
         return f'Line({self._p1}, {self._p2})'
 
-    @property
-    def direction(self):
-        return self._p2 - self._p1
-
     def point(self, factor=0):
         """
-        :return: point on the line based on the first initialization point added
-        to the factor-th of the vector from first to second initialization point.
+        :return: point on the line based on the first initialization
+        point added to the factor-th of the direction of the line.
         """
-        return self._p1 + factor*self.direction
-
-    def move(self, vec: Vector):
-        """:return: line moved with given vector"""
-        return Line(self._p1 + vec, self._p2 + vec)
+        raise NotImplementedError
 
     def project(self, point: Vector):
-        """:return: projection of given point onto the line"""
-        # move line and argument such that both start in (0, 0)
-        origin_line = self.move(-self._p1)
-        span = origin_line._p2 - origin_line._p1
-        point = point - self._p1
-        return ((point*span)/(span*span))*span + self._p1
+        """
+        :return: projection of given point onto the line such that the distance
+        between the original point and the projected point is minimal.
+        """
+        raise NotImplementedError
 
-    def orthogonal_line(self, through: Vector):
-        """:return: Line orthogonal of self through the given point"""
-        span = self._p2 - self._p1
-        return Line(through, through + span.orthogonal_vector())
+    def orthogonal(self, through: Vector):
+        """:return: line orthogonal to self and through the given point"""
+        raise NotImplementedError
 
     def intersection(self, other):
         """
-        :return: intersection point of self and other or None if both lines don't
-        intersect or a line of both lines share an infinite number of points
+        :param other: instance of StraightLine of identical type like self
+        :return: intersection point of self and other or None if both lines
+        don't intersect or self if both lines share an infinite number of points
         """
-        if self == other:
-            return self
-        if self.direction.parallel_to(other.direction):
-            return None
-
-        # solve the linear equation system derived from equality of both lines
-        o_dir = other.direction
-        s_dir = self.direction
-        factor = (o_dir.y*(self.point(0).x - other.point(0).x) \
-                  - o_dir.x*(self.point(0).y - other.point(0).y)) \
-            / (o_dir.x*s_dir.y - o_dir.y*s_dir.x)
-        # divisor is the determinant of line directions and != 0 if
-        # lines aren't parallel (which is guaranteed here)
-        return self.point(factor)
+        raise NotImplementedError
