+++ tests/test_parse.py
@@ -5,7 +5,7 @@ import tempfile
 from gpxpy.gpx import GPX, GPXTrackPoint, GPXXMLSyntaxException
 
 from linesman.parse import lonlat_str, lonlat_pair_str, gpx_file, gpx_extract_points
-from linesman.util import Vector, Line
+from linesman.geometry import Vector, Line
 
 
 def test_lonlat_str_no_comma():
